+++ Code/Combat/hud.cpp
@@ -256,7 +256,7 @@ int							HUDHelpTextState = HUD_HELP_TEXT_DISPLAYING;
 /*
 **
 */
-unsigned long COLOR( float alpha, unsigned long color = 0x00FFFFFF )
+uint32_t COLOR( float alpha, uint32_t color = 0x00FFFFFF )
 {
 	alpha = WWMath::Clamp( alpha, 0, 1 ) * 255.0f;
 	color &= 0x00FFFFFF;
@@ -264,7 +264,7 @@ unsigned long COLOR( float alpha, unsigned long color = 0x00FFFFFF )
 	return color;
 }
 
-unsigned long Get_Health_Color( float percent )
+uint32_t Get_Health_Color( float percent )
 {
 	Vector3	color = HUDGlobalSettingsDef::Get_Instance()->Get_Health_High_Color();
 	if ( percent <= 0.5f ) {
@@ -1150,7 +1150,7 @@ static	void	Weapon_Chart_Update( void )
 					continue;
 				}
 
-				long color = 0x000FF00;		// Dim
+				uint32_t color = 0x000FF00;		// Dim
 				if ( weapon_bag->Get_Index() == i ) {
 					color = 0x0000FF00;			// Bright
 				}
@@ -1259,8 +1259,8 @@ static	void	Damage_Add_Indicator( int index, float start_x, float start_y, float
 	}
 	uv.Scale( INFO_UV_SCALE );
 
-	unsigned long color_bits = (int)(DamageIndicatorIntensity[index] * 255) & 0x000000FF;
-  	unsigned long color = color_bits | color_bits<<8 | color_bits<<16;
+	uint32_t color_bits = (int)(DamageIndicatorIntensity[index] * 255) & 0x000000FF;
+  	uint32_t color = color_bits | color_bits<<8 | color_bits<<16;
 
 	switch (index) {
 		case 3:
@@ -1411,7 +1411,7 @@ static	void	Target_Shutdown( void )
 		bool	InfoDebug = false;
 
 static RectClass Get_Target_Box( PhysicalGameObj * obj );
-static void	Target_Box_Edge( const Vector2 & a, const Vector2 & b, unsigned int color );
+static void	Target_Box_Edge( const Vector2 & a, const Vector2 & b, uint32_t color );
 
 static	void	Target_Update( void )
 {
@@ -1787,7 +1787,7 @@ static RectClass Get_Target_Box( PhysicalGameObj * obj )
 	return info_box;
 }
 
-static void	Target_Box_Edge( const Vector2 & a, const Vector2 & b, unsigned int color )
+static void	Target_Box_Edge( const Vector2 & a, const Vector2 & b, uint32_t color )
 {
 	float percent = 0.2f;
 	Vector2 a_ = b - a;
@@ -2005,7 +2005,7 @@ static	void	Objective_Update( void )
 						if ( ObjectiveManager::Get_Objective(index) != NULL ) {
 							color3 = ObjectiveManager::Get_Objective(index)->Type_To_Color();
 						}
-						unsigned int color = color3.Convert_To_ARGB();
+						uint32_t color = color3.Convert_To_ARGB();
 						renderer->Add_Quad( star_box, color  );
 						ObjectivePogRenderers.Add( renderer );
 					}
@@ -2430,7 +2430,7 @@ static	void	Info_Update_Health_Shield( void )
 		health = 1;
 	}
 	//text.Format( "%03d", (int)health );
-	long lhealth=WWMath::Float_To_Long(health);
+	int32_t lhealth=WWMath::Float_To_Long(health);
 	WCHAR tmp_text[5];
 	Generate_WChar_Text_From_Number(tmp_text,4,3,lhealth);
 
@@ -2502,7 +2502,7 @@ static	void	Info_Update_Health_Shield( void )
 		InfoShieldCountRenderer->Reset();
 //		StringClass	text;
 //		text.Format( "%03d", (int)shield );
-		long lshield=WWMath::Float_To_Long(shield);
+		int32_t lshield=WWMath::Float_To_Long(shield);
 		WCHAR tmp_text[5];
 		Generate_WChar_Text_From_Number(tmp_text,4,3,lshield);
 		InfoShieldCountRenderer->Set_Location( draw.Upper_Left() + Vector2( 4,4) );
@@ -2928,7 +2928,7 @@ void 	HUDClass::Think()
 		//
 		float width					= status_bar_rect.Width() * HUDInfo::Get_Action_Status_Value();
 		status_bar_rect.Right	= status_bar_rect.Left + int(width);
-		DWORD color					= VRGB_TO_INT32( Vector3( 1.0F, 0.25F, 0 ) );
+		uint32_t color					= VRGB_TO_INT32( Vector3( 1.0F, 0.25F, 0 ) );
 		RenderImages[ACTION_STATUSBAR_RENDERER]->Add_Quad( status_bar_rect, color );
 
 	} else {
@@ -2967,7 +2967,7 @@ void 	HUDClass::Think()
 	Vector2	radar_center = InfoBase + RADAR_CENTER_OFFSET;
 	RadarManager::Update( tm, radar_center );
 
-	unsigned long reticle_color = HUDGlobalSettingsDef::Get_Instance()->Get_No_Relation_Color().Convert_To_ARGB();
+	uint32_t reticle_color = HUDGlobalSettingsDef::Get_Instance()->Get_No_Relation_Color().Convert_To_ARGB();
 
 	if ( HUDInfo::Get_Weapon_Target_Object() != NULL ) {
 		reticle_color = HUDGlobalSettingsDef::Get_Instance()->Get_Friendly_Color().Convert_To_ARGB();
