+++ Code/Commando/CDKeyAuth.cpp
@@ -111,7 +111,7 @@ void CCDKeyAuth::DisconnectUser(int localid) {
 
 }
 
-void CCDKeyAuth::AuthenticateUser(int localid, ULONG ip, char *challenge, char *authstring) {
+void CCDKeyAuth::AuthenticateUser(int localid, uint32_t ip, char *challenge, char *authstring) {
 
 	// Customize this with our playerdata struct
 	// Take the response from our challenge that we sent to the client
@@ -147,7 +147,7 @@ void CCDKeyAuth::AuthSerial(const char *challenge, StringClass &resp) {
 	*outb = 0;
 
 	// MD5 Hash Here.
-	MD5Digest((BYTE *)cdkey, strlen(cdkey), md5hash);
+	MD5Digest((uint8_t *)cdkey, strlen(cdkey), md5hash);
 
 	// hashserial, challenge, outbuf
 	gcd_compute_response(md5hash, challenge, response);
